requirement.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.16.25
botocore==1.19.25
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.5.0
django-countries==6.1.3
django-rest-knox==4.1.0
djangorestframework==3.12.2
dynaconf==3.1.2
hvac==0.10.5
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.11
loguru==0.5.3
MarkupSafe==1.1.1
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
s3transfer==0.3.3
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
validators==0.18.1
zope.interface==5.2.0
